Update clonetodevice
I had set a variable containing the number of partitions found on the target device but never made use of it.
The script is written to offer a "refresh" only if the target has the proper number and type of partitions.
Any number of partitions on the target other than two now offers to format.
I discovered this after adding a third partition to the end of my slave drive and then trying to clone. It offered to refresh an existing clone, which it should not have. But since the first two partitions were correct, it worked fine. But that was just blind luck. :)