Rename clonetodevice to mirror2device
Although I think of this as a cloner, (appearance-wise, it is) I realize it's not a true bit to bit drive clone. So... I'll change the name to mirror2device.

Actually, with SSDs and wear leveling, does even dd produce a true bit to bit clone? Maybe, maybe not, suppose it doesn't matter. :)