Update mirror2device
Added a 10 second sleep after sync command in the function that updates grub. Dismounting target device failed a couple times due to "target device busy". I hadn't seen that before, I though the sync command would be enough but apparently not. It was not anything to cause the mirror to fail but leaving the target mounted, with a warning, seemed untidy. I haven't seen it since but time will tell.