Add files via upload
Special instructions for using this utility to remove all Windows partitions. Not really why I wrote this but I used it this way once and I've seen numerous other folks asking how to remove Windows partitions and resize the Linux partitions. There's other ways to do that but this is so much easier and handles the size difference without jumping through hoops. Still... use at your own risk. Good luck if you try. I'm not a programmer, just like to be helpful when I can. :)